feat(core): add installOpenapiValidationMiddleware
This method allows us to re-use the API server's internal
mechanisms to configure the OpenAPI spec validation in
test cases where we cannot depend on the API server itself
due to circular dependencies.
So this method is designed to be used both by the API server
and the test cases at the same time.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>